+++ Tests/IntegrationTests.Shared/InstanceTests.cs
@@ -189,5 +189,70 @@ namespace IntegrationTests
             // Arrange
             Assert.Throws<RealmPermissionDeniedException>(() => Realm.GetInstance("/"));
         }
+
+
+        class LoneClass : RealmObject
+        {
+            public string Name { get; set;}
+        }
+
+        [Test]
+        public void RealmWithOneClassWritesDesiredClass()
+        {
+            // Arrange
+            var config = new RealmConfiguration("RealmWithOneClass.realm");
+            Realm.DeleteRealm(config);
+            config.ObjectClasses = new Type[] {typeof(LoneClass)};
+
+            // Act
+            using (var lonelyRealm = Realm.GetInstance(config)) {
+                lonelyRealm.Write( () => {
+                    var p = lonelyRealm.CreateObject<LoneClass>();
+                    p.Name = "The Singular";
+                }); 
+
+                // Assert
+                Assert.That(lonelyRealm.All<LoneClass>().Count(), Is.EqualTo(1));
+            }
+        }
+
+        [Test]
+        public void RealmWithOneClassThrowsIfUseOther()
+        {
+            // Arrange
+            var config = new RealmConfiguration("RealmWithOneClass.realm");
+            Realm.DeleteRealm(config);
+            config.ObjectClasses = new Type[] {typeof(LoneClass)};
+
+            // Act and assert
+            using (var lonelyRealm = Realm.GetInstance(config)) {
+                using (var trans = lonelyRealm.BeginWrite())
+                {
+                    Assert.Throws<ArgumentException>(() =>
+                        {
+                            lonelyRealm.CreateObject<Person>(); 
+                        }, 
+                        "Can't create an object with a class not included in this Realm"); 
+                } // transaction
+            }  // realm
+        }
+
+
+        [Test]
+        public void RealmObjectClassesOnlyAllowRealmObjects()
+        {
+            // Arrange
+            var config = new RealmConfiguration("RealmWithOneClass.realm");
+            Realm.DeleteRealm(config);
+            config.ObjectClasses = new Type[] {typeof(LoneClass), typeof(object)};
+
+            // Act and assert
+            Assert.Throws<ArgumentException>(() =>
+                {
+                    Realm.GetInstance(config); 
+                }, 
+                "Can't have classes in the list which are not RealmObjects"); 
+        }
+
     }
 }